refactor: move RunnableAdd to operator module and update imports
